feat: update diagramStyles.js
